etl/requirements.txt
dbt-core==1.8.5
dbt-postgres==1.8.2
psycopg2-binary==2.9.10
